ENH: RDS postgres instance and custom resource (#48)
RDS instance is a T2 micro (in AWS free tier) and uses postgres 12.8 since this was also required to get access to free tier. Set it to do no backups and credentials are from secretmanager secret. Security group is given that allows all external IPV4 access and all internal on port 5432.

Decided to run the RDS instance in a public VPC subnet, which is run in two AZs (this is a requirement for RDS deployments). In future for security we may want to move this to a private / isolated subnet- avoiding for now since NAT gateway prices. Since its in a public subnet there is no need currently to put lambdas etc in the same VPC.

Running pgadmin for a web UI to interact with postgres. Run with `docker run -p 80:80 -e 'PGADMIN_DEFAULT_EMAIL=admin@admin.com' -e 'PGADMIN_DEFAULT_PASSWORD=admin' dpage/pgadmin4:latest`.

Added a custom resource lambda (init-db) that connects to the postgres RDS instance and creates a database for the web app as well as creating the tickers and holdings tables. Resource seems to run on create and delete, and only truely fails on exceptions being raised. In the meantime when updating, can just run a manual event through the AWS web UI.