ENH: compute units and gain & snackbar (#56)
Changes to the holding list view:
- Now pulls transactions and uses them to determine units / gains.
- Column widths in table are explicitly set.

Fixed the holding price chart to correctly show a year's worth of data (instead of the last 365 data items).

Changed the add transaction dialog to show a snackbar (toast) on completion and now updates state of the transactions in react. In order to do this the add transaction lambda needed to include a copy of the inserted transaction in the API response. Changed secondary theme colour from yellow to blue.