ENH: start of cognito integration & state updates when deleting / adding (#57)
Started cognito integration- added a UserPool to the frontend stack and a component which allows login & viewing session state. Still need to find a way to relate to DB account ID (federated cognito?). Using a cognito-hosted domain currently.

State of transactions and holdings are now updated when a delete or an add is completed successfully. In order to update state for an add, the lambdas had to be updated to return the JSON representation of the added DB item.

Other changes:
- VS code settings to exclude .git and node_modules files from watch.
- Added delete holding lambda.
- Removed units and market value from holdings table & queries since it is now computed in FE.
- Updated snackbar to allow it to be provided with the severity and message.