ENH: profile identicons & logout and S3 `autoDeleteObjects` (#60)
Adds identicon (similar to github default profile picture) which are created based on the AWS cognito account "sub" ID. Also adds a drop down menu when the identicon is clicked which gives an option to logout. 

Set `autoDeleteObjects` property on S3 bucket in CDK so that it can be deleted along with the rest of the stack without needing to manually empty first.