ENH: price per unit in new transaction form and fix cloudfront refresh error (#61)
- Closes #46 - added a field for price per unit and when either field is set the other is updated.
- Closes #42 - added a custom error response where any 404 errors are redirected to 200 errors, and instead the react SPA will be responsible for 404 pages.

Updated insert holding to include an account ID and removed units which is now calculated based on transactions rather than persisted. Also updated certain lambdas to require the account ID to be passed.

Adds authentication checks for certain pages, with redirects to /login if not authenticated. Removed column width specification in holdings table.